+++ ui/src/pages/namespace/Explorer/Page/poc/PageCompiler/context/utils/useBlockTypes.tsx
@@ -1,15 +1,9 @@
 import {
-  AllBlocksType,
-  FormBlocks,
-  InlineBlocksType,
-} from "../../../schema/blocks";
-import {
   Captions,
   Columns2,
   Database,
   Heading1,
   Image,
-  LucideIcon,
   RectangleHorizontal,
   Repeat2,
   Table,
@@ -17,10 +11,8 @@ import {
   TextCursorInput,
 } from "lucide-react";
 
-import { BlockEditFormProps } from "../../../BlockEditor";
-import { BlockPathType } from "../../Block";
+import { BlockTypeConfig } from "./types";
 import { Dialog as DialogForm } from "../../../BlockEditor/Dialog";
-import { DirektivPagesType } from "../../../schema";
 import { Form as FormForm } from "../../../BlockEditor/Form";
 import { Headline } from "../../../BlockEditor/Headline";
 import { Image as ImageForm } from "../../../BlockEditor/Image";
@@ -30,34 +22,6 @@ import { Table as TableForm } from "../../../BlockEditor/Table";
 import { Text as TextForm } from "../../../BlockEditor/Text";
 import { findAncestor } from ".";
 import { useTranslation } from "react-i18next";
-
-// common types for block type configuration
-type BlockTypeConfigBase = {
-  label: string;
-  icon: LucideIcon;
-  allow: (page: DirektivPagesType, path: BlockPathType) => boolean;
-};
-
-// inline blocks don't have a form component
-type BlockTypeConfigWithoutForm = {
-  [K in InlineBlocksType]: {
-    type: K;
-    formComponent?: never;
-    defaultValues: Extract<AllBlocksType, { type: K }>;
-  };
-}[InlineBlocksType];
-
-// blocks that require a form, must have a form component that implments a form for that very block type
-type BlockTypeConfigWithForm = {
-  [K in FormBlocks as K["type"]]: {
-    type: K["type"];
-    formComponent: React.ComponentType<BlockEditFormProps<K>>;
-    defaultValues: K;
-  };
-}[FormBlocks["type"]];
-
-type BlockTypeConfig = BlockTypeConfigBase &
-  (BlockTypeConfigWithForm | BlockTypeConfigWithoutForm);
 
 export const useBlockTypes = (): BlockTypeConfig[] => {
   const { t } = useTranslation();
